refactor: separate CafeDetailSectionDetail
